refactor: repositorio generico (EditarAsync, RecuperarAsync por predicado, ListarAsync paginado/ordenado), Id int, ListarProdutosRequest, TipoOrdenacao

+++ src/MinhaApi.Domain/Produtos/Commands/AtualizarProdutoCommand.cs
@@ -2,4 +2,4 @@ namespace MinhaApi.Domain.Produtos.Commands;
 
 // Carrega o Id junto (diferente do Request, que recebe o Id separado via rota) -
 // um Command representa a intencao COMPLETA de uma acao, entao fica autocontido.
-public record AtualizarProdutoCommand(long Id, string Nome, decimal Preco);
+public record AtualizarProdutoCommand(int Id, string Nome, decimal Preco);
